+++ crypto/openpgp/constants.go
@@ -1,0 +1,5 @@
+package openpgp
+
+const (
+	ENV_MASTER_GPG_ID_KEY = "TROUSSEAU_MASTER_GPG_ID"
+)