+++ remote/gist/constants.go
@@ -1,0 +1,3 @@
+package gist
+
+const TOKEN="9c8f97e57b9a5ea6dae59e9762c4deb801fc6b91"