+++ trousseau/actions.go
@@ -3,6 +3,7 @@ package trousseau
 import (
 	"fmt"
 	"github.com/codegangsta/cli"
+	"github.com/oleiade/trousseau/crypto"
 	"github.com/oleiade/trousseau/dsn"
 	"io"
 	"log"
@@ -11,30 +12,18 @@ import (
 	"time"
 )
 
-// hasExpectedArgs checks whether the number of args are as expected.
-func hasExpectedArgs(args []string, expected int) bool {
-	switch expected {
-	case -1:
-		if len(args) > 0 {
-			return true
-		} else {
-			return false
-		}
-	default:
-		if len(args) == expected {
-			return true
-		} else {
-			return false
-		}
-	}
-}
-
 func CreateAction(c *cli.Context) {
 	if !hasExpectedArgs(c.Args(), 1) {
 		log.Fatal("Incorrect number of arguments to 'configure' command")
 	}
 
 	recipients := strings.Split(c.Args()[0], ",")
+
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+		Recipients: recipients,
+	}
 
 	meta := Meta{
 		CreatedAt:        time.Now().String(),
@@ -44,9 +33,12 @@ func CreateAction(c *cli.Context) {
 	}
 
 	// Create and write empty store file
-	CreateStoreFile(gStorePath, &meta)
+	err := CreateStoreFile(gStorePath, opts, &meta)
+	if err != nil {
+		log.Fatal(err)
+	}
 
-	fmt.Println("trousseau created")
+	fmt.Println("Trousseau data store succesfully created")
 }
 
 func PushAction(c *cli.Context) {
@@ -70,6 +62,7 @@ func PushAction(c *cli.Context) {
 		if err != nil {
 			log.Fatal(err)
 		}
+		fmt.Printf("Trousseau data store succesfully pushed to s3")
 	case "scp":
 		privateKey := c.String("ssh-private-key")
 
@@ -78,10 +71,22 @@ func PushAction(c *cli.Context) {
 			log.Fatal(err)
 		}
 
+		if c.Bool("ask-password") == true {
+			password := PromptForPassword()
+			endpointDsn.Secret = password
+		}
+
 		err = uploadUsingScp(endpointDsn, privateKey)
 		if err != nil {
 			log.Fatal(err)
 		}
+		fmt.Printf("Trousseau data store succesfully pushed to ssh remote storage")
+	case "gist":
+		err = uploadUsingGist(endpointDsn)
+		if err != nil {
+			log.Fatal(err)
+		}
+		fmt.Printf("Trousseau data store succesfully pushed to gist")
 	}
 }
 
@@ -106,6 +111,7 @@ func PullAction(c *cli.Context) {
 		if err != nil {
 			log.Fatal(err)
 		}
+		fmt.Printf("Trousseau data store succesfully pulled from S3")
 	case "scp":
 		privateKey := c.String("ssh-private-key")
 
@@ -114,11 +120,31 @@ func PullAction(c *cli.Context) {
 			log.Fatal(err)
 		}
 
+		if c.Bool("ask-password") == true {
+			password := PromptForPassword()
+			endpointDsn.Secret = password
+		}
+
 		err = DownloadUsingScp(endpointDsn, privateKey)
 		if err != nil {
 			log.Fatal(err)
 		}
+		fmt.Printf("Trousseau data store succesfully pulled from ssh remote storage")
+	case "gist":
+		err = DownloadUsingGist(endpointDsn)
+		if err != nil {
+			log.Fatal(err)
+		}
+		fmt.Printf("Trousseau data store succesfully pulled from gist")
+	default:
+		if endpointDsn.Scheme == "" {
+			log.Fatalf("No dsn scheme supplied")
+		} else {
+			log.Fatalf("Invalid dsn scheme supplied: %s", endpointDsn.Scheme)
+		}
 	}
+
+	fmt.Printf("Trousseau data store succesfully pulled from remote storage")
 }
 
 func ExportAction(c *cli.Context) {
@@ -147,7 +173,7 @@ func ExportAction(c *cli.Context) {
 		log.Fatal(err)
 	}
 
-	fmt.Println("trousseau exported")
+	fmt.Printf("Trousseau data store exported to %s", outputFilePath)
 }
 
 func ImportAction(c *cli.Context) {
@@ -156,40 +182,43 @@ func ImportAction(c *cli.Context) {
 	}
 
 	var err error
-	var inputFilePath string = c.Args()[0]
-	var outputFilePath string = gStorePath
-	var inputFile *os.File
-	var outputFile *os.File
+	var importedFilePath string = c.Args()[0]
+	var localFilePath string = gStorePath
+	var strategy *ImportStrategy = new(ImportStrategy)
 
-	inputFile, err = os.Open(inputFilePath)
-	defer inputFile.Close()
+	// Transform provided merging startegy flags
+	// into a proper ImportStrategy byte.
+	err = strategy.FromCliContext(c)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	// If trousseau data store does not exist create
-	// it through the import process, otherwise just
-	// override it's content
-	if !pathExists(outputFilePath) {
-		outputFile, err = os.Create(outputFilePath)
-		defer outputFile.Close()
-		if err != nil {
-			log.Fatal(err)
-		}
-	} else {
-		outputFile, err = os.OpenFile(outputFilePath, os.O_WRONLY, 0744)
-		defer outputFile.Close()
-		if err != nil {
-			log.Fatal(err)
-		}
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
 	}
 
-	_, err = io.Copy(outputFile, inputFile)
+	localStore, err := LoadStore(localFilePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	fmt.Println("trousseau imported")
+	importedStore, err := LoadStore(importedFilePath, opts)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	err = ImportStore(importedStore, localStore, *strategy)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	err = localStore.Sync()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Println("Trousseau data store imported")
 }
 
 func AddRecipientAction(c *cli.Context) {
@@ -199,29 +228,24 @@ func AddRecipientAction(c *cli.Context) {
 
 	recipient := c.Args()[0]
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
-	if err != nil {
-		log.Fatal(err)
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
 	}
 
-	err = store.Decrypt()
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
 
 	err = store.DataStore.Meta.AddRecipient(recipient)
 
-	err = store.Encrypt()
+	err = store.Sync()
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	err = store.WriteToFile(gStorePath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Printf("%s added to trousseau recipients", recipient)
+	fmt.Printf("%s recipient added to trousseau data store", recipient)
 }
 
 func RemoveRecipientAction(c *cli.Context) {
@@ -231,29 +255,24 @@ func RemoveRecipientAction(c *cli.Context) {
 
 	recipient := c.Args()[0]
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	err = store.Decrypt()
+	err = store.Meta.RemoveRecipient(recipient)
+
+	err = store.Sync()
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	err = store.DataStore.Meta.RemoveRecipient(recipient)
-
-	err = store.Encrypt()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = store.WriteToFile(gStorePath)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Printf("%s removed from trousseau recipients", recipient)
+	fmt.Printf("%s recipient removed from trousseau data store", recipient)
 }
 
 func GetAction(c *cli.Context) {
@@ -261,7 +280,12 @@ func GetAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'get' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -279,7 +303,12 @@ func SetAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'set' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -289,7 +318,7 @@ func SetAction(c *cli.Context) {
 		log.Fatal(err)
 	}
 
-	err = store.WriteToFile(gStorePath)
+	err = store.Sync()
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -302,7 +331,12 @@ func DelAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'del' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -312,7 +346,7 @@ func DelAction(c *cli.Context) {
 		log.Fatal(err)
 	}
 
-	err = store.WriteToFile(gStorePath)
+	err = store.Sync()
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -325,7 +359,12 @@ func KeysAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'keys' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -345,7 +384,12 @@ func ShowAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'show' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -365,17 +409,40 @@ func MetaAction(c *cli.Context) {
 		log.Fatal("Incorrect number of arguments to 'meta' command")
 	}
 
-	store, err := NewEncryptedStoreFromFile(gStorePath, c.GlobalString("passphrase"))
+	opts := &crypto.Options{
+		Algorithm:  crypto.GPG_ENCRYPTION,
+		Passphrase: gPasshphrase,
+	}
+
+	store, err := LoadStore(gStorePath, opts)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	pairs, err := store.Meta()
+	pairs, err := store.Metadata()
 	if err != nil {
 		log.Fatal(err)
 	}
 
 	for _, pair := range pairs {
 		fmt.Printf("%s: %s\n", pair.Key, pair.Value)
+	}
+}
+
+// hasExpectedArgs checks whether the number of args are as expected.
+func hasExpectedArgs(args []string, expected int) bool {
+	switch expected {
+	case -1:
+		if len(args) > 0 {
+			return true
+		} else {
+			return false
+		}
+	default:
+		if len(args) == expected {
+			return true
+		} else {
+			return false
+		}
 	}
 }