+++ trousseau/cli_helpers.go
@@ -1,0 +1,11 @@
+package trousseau
+
+import (
+	"fmt"
+	"github.com/howeyc/gopass"
+)
+
+func PromptForPassword() string {
+	fmt.Printf("Password: ")
+	return string(gopass.GetPasswd())
+}