+++ trousseau/constants.go
@@ -1,6 +1,6 @@
 package trousseau
 
-const TROUSSEAU_VERSION = "0.2.0"
+const TROUSSEAU_VERSION = "0.2.4"
 
 const STORE_FILENAME = ".trousseau"
 
@@ -15,4 +15,11 @@ const (
 	ENV_KEYRING_USER_KEY    = "USER"
 	ENV_MASTER_GPG_ID_KEY   = "TROUSSEAU_MASTER_GPG_ID"
 	ENV_SSH_PRIVATE_KEY     = "TROUSSEAU_PRIVATE_KEY"
+)
+
+// Import strategies enumeration
+const (
+	IMPORT_YOURS     = 0x0
+	IMPORT_THEIRS    = 0x1
+	IMPORT_OVERWRITE = 0x2
 )