+++ trousseau/globals.go
@@ -8,12 +8,9 @@ import (
 // Global data store file path
 var gStorePath string = filepath.Join(os.Getenv("HOME"), STORE_FILENAME)
 
-// Gnupg keyrings files
-var gPubringFile string = filepath.Join(os.Getenv("HOME"), ".gnupg", "pubring.gpg")
-var gSecringFile string = filepath.Join(os.Getenv("HOME"), ".gnupg", "secring.gpg")
-
 // Gnupg trousseau master gpg key id
 var gMasterGpgId string = os.Getenv(ENV_MASTER_GPG_ID_KEY)
+var gPasshphrase string = GetPassphrase()
 
 // Ssh default identity file path
 var gPrivateRsaKeyPath string = filepath.Join(os.Getenv("HOME"), ".ssh", "id_rsa")